refactor: 1단계 기능 테스트 추가 및 리팩토링
