+++ src/main/java/coupon/repository/CouponRepository.java
@@ -3,7 +3,7 @@ package coupon.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import coupon.repository.CouponEntity;
+import coupon.repository.entity.CouponEntity;
 
 @Repository
 public interface CouponRepository extends JpaRepository<CouponEntity, Long> {
