+++ src/main/java/coupon/repository/entity/MemberEntity.java
@@ -1,4 +1,4 @@
-package coupon.repository;
+package coupon.repository.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
